Crate melvin

Melvin is a library for configuring logical volumes in the style of LVM Melvin is not compatible with LVM.

Modules

dm
Low-level devicemapper configuration of the running kernel.

parser
Parsing MLV's text-based configuration format.

Structs

Device
A struct containing the device's major and minor numbers

LV
A Logical Volume that is created from a Volume Group.

PV
A Physical Volume that is part of a Volume Group.

PvHeader
A block device that has been initialized to be a MLV Physical Volume, but that may not be part of a VG yet.

VG
A Volume Group allows multiple Physical Volumes to be treated as a storage pool that can then be used to allocate Logical Volumes.

Functions

pvheader_scan
Scan a list of directories for block devices containing MLV PV labels.

vg_list
Query lvmetad for a list of Volume Groups on the system.