Struct melvin::dm::DM

pub struct DM<'a> {
    // some fields omitted
}

Context needed for communicating with devicemapper.

Methods

impl<'a> DM<'a>

fn new(vg: &'a VG) -> Result<Self>

Create a new context for communicating about a given VG with DM.

fn dm_majors() -> BTreeSet<u32>

Determine if a major number is a major number used by DM.

fn get_version(&self) -> Result<(u32, u32, u32)>

Devicemapper version information: Major, Minor, and patchlevel versions.

fn list_devices(&self) -> Result<Vec<(String, Device)>>

Returns a list of tuples containing DM device names within this VG, and their major/minor device numbers.

fn list_deps(&self, dev: Device) -> Result<Vec<Device>>

Query DM for which devices depend on this device.

fn activate_device(&self, lv: &mut LV) -> Result<()>

Activate a Logical Volume.

Also populates the LV's device field.

fn deactivate_device(&self, lv: &mut LV) -> Result<()>

Deactivate a Logical Volume.