Module melvin::parser

Parsing MLV's text-based configuration format.

Enums

Entry
Each value in an LvmTextMap is an Entry.

Traits

TextMapOps
Operations that can be used to extract values from an LvmTextMap.

Functions

buf_to_textmap
Generate an LvmTextMap from a textual MLV configuration string.

status_from_textmap
Status may be either a string or a list of strings. Convert either into a list of strings.

textmap_to_buf
Generate a textual MLV configuration string from an LvmTextMap.

Type Definitions

LvmTextMap
A Map that represents MLV metadata.