Struct melvin::pvlabel::MDA

pub struct MDA {
    // some fields omitted
}

A handle to an LVM on-disk metadata area (MDA)

Methods

impl MDA

fn new(path: &Path) -> Result<MDA>

Construct an MDA given a path to a block device containing an LVM Physical Volume (PV)

fn read_metadata(&mut self) -> Result<LvmTextMap>

Read the metadata contained in the metadata area.

fn write_metadata(&mut self, map: &LvmTextMap) -> Result<()>

Write a new version of the metadata to the metadata area.