Struct melvin::pvlabel::PvHeader

pub struct PvHeader {
    pub uuid: String,
    pub size: u64,
    pub ext_version: u32,
    pub ext_flags: u32,
    pub data_areas: Vec<PvArea>,
    pub metadata_areas: Vec<PvArea>,
    pub bootloader_areas: Vec<PvArea>,
    pub dev_path: PathBuf,
}

A struct containing the values in the PV header. It contains pointers to the data area, and possibly metadata areas and bootloader area.

Fields

uuid

The unique identifier.

size

Size in bytes of the entire PV

ext_version

Extension version. If 1, we look for an extension header that may contain a reference to a bootloader area.

ext_flags

Extension flags, of which there are none.

data_areas

A list of the data areas.

metadata_areas

A list of the metadata areas.

bootloader_areas

A list of the bootloader areas.

dev_path

The device this pvheader is for

Methods

impl PvHeader

fn find_in_dev(path: &Path) -> Result<PvHeader>

Find the PvHeader struct in a given device.

fn initialize(path: &Path) -> Result<PvHeader>

Initialize a device as a PV with reasonable defaults: two metadata areas, no bootsector area, and size based on the device's size.

fn read_metadata(&self) -> Result<LvmTextMap>

Read the metadata contained in the metadata area. In the case of multiple metadata areas, return the information from the first valid one.

fn write_metadata(&mut self, map: &LvmTextMap) -> Result<()>

Write the given metadata to all active metadata areas in the PV.