Struct melvin::PvHeader

pub struct PvHeader {
    pub uuid: String,
    pub size: u64,
    pub data_areas: Vec<PvArea>,
    pub metadata_areas: Vec<PvArea>,
    pub bootloader_areas: Vec<PvArea>,
    pub dev_path: PathBuf,
    // some fields omitted
}

A block device that has been initialized to be a LVM Physical Volume, but that may not be part of a VG yet.

Fields

uuid

The unique identifier.

size

Size in bytes of the entire PV.

data_areas

A list of the data areas.

metadata_areas

A list of the metadata areas.

bootloader_areas

A list of the bootloader areas.

dev_path

The path to the device this pvheader is within.

Methods

impl PvHeader

fn find_in_dev(path: &Path) -> Result<PvHeader>

Find the PvHeader struct in a given device.

fn initialize(path: &Path) -> Result<PvHeader>

Initialize a device as a PV with reasonable defaults: two metadata areas, no bootsector area, and size based on the device's size.

fn read_metadata(&self) -> Result<LvmTextMap>

Read the metadata contained in the metadata area. In the case of multiple metadata areas, return the information from the first valid one.

fn write_metadata(&mut self, map: &LvmTextMap) -> Result<()>

Write the given metadata to all active metadata areas in the PV.