Struct melvin::VG

pub struct VG {
    // some fields omitted
}

A Volume Group allows multiple Physical Volumes to be treated as a storage pool that can then be used to allocate Logical Volumes.

Methods

impl VG

fn create(name: &str, pv_paths: Vec<&Path>) -> Result<VG>

Create a Volume Group from one or more PVs.

fn pv_add(&mut self, path: &Path) -> Result<()>

Add a non-affiliated PV to this VG.

fn pv_remove(&mut self, pvh: &PvHeader) -> Result<()>

Remove a PV. It must be unused by any LVs.

fn lv_create_linear(&mut self, name: &str, extent_size: u64) -> Result<()>

Create a new linear logical volume in the volume group.

fn lv_create_thinpool(&mut self, name: &str, thin_meta: &str, thin_data: &str) -> Result<()>

Create a thin pool from existing metadata and data volumes. These will be renamed to "_tmeta" and "_tdata". In addition, a spare metadata volume will be created if one does not already exist.

See the kernel's thin-provisioning.txt for the exact calculation, but a reasonable size for the metadata volume (assuming default thinpool chunk size of 64KiB) is 1/1000 the data volume, minimum 2MiB.

fn lv_remove(&mut self, name: &str) -> Result<()>

Destroy a logical volume.

fn extents_in_use(&self) -> u64

The total number of extents in use in the volume group.

fn extents_free(&self) -> u64

The total number of free extents in the volume group.

fn extents(&self) -> u64

The total number of extents in the volume group.

fn pv_list(&self) -> Vec<Device>

Returns a list of PV Devices that make up the VG.

fn pv_get(&self, dev: Device) -> Option<&PV>

Returns a reference to the PV matching the Device.

fn lv_list(&self) -> Vec<String>

Returns a list of the names of LVs in the VG.

fn lv_get(&self, name: &str) -> Option<&LV>

Returns a reference to the LV matching the name.

fn name(&self) -> &str

Returns the name of the VG.

fn id(&self) -> &str

Returns the UUID of the VG.

fn extent_size(&self) -> u64

Returns how many 512-byte sectors make up each extent in the VG.