Struct melvin::vg::VG

pub struct VG {
    pub name: String,
    pub id: String,
    pub seqno: u64,
    pub format: String,
    pub status: Vec<String>,
    pub flags: Vec<String>,
    pub extent_size: u64,
    pub max_lv: u64,
    pub max_pv: u64,
    pub metadata_copies: u64,
    pub pvs: BTreeMap<String, PV>,
    pub lvs: BTreeMap<String, LV>,
}

A Volume Group.

Fields

name

Name.

id

Uuid.

seqno

The generation of metadata this VG represents.

format

Always "LVM2".

status

Status.

flags

Flags.

extent_size

Size of each extent, in 512-byte sectors.

max_lv

Maximum number of LVs, 0 means no limit.

max_pv

Maximum number of PVs, 0 means no limit.

metadata_copies

How many metadata copies (?)

pvs

Physical Volumes within this volume group.

lvs

Logical Volumes within this volume group.

Methods

impl VG

fn extents_in_use(&self) -> u64

The total number of extents in use in the volume group.

fn extents_free(&self) -> u64

The total number of free extents in the volume group.

fn extents(&self) -> u64

The total number of extents in the volume group.

fn add_pv(&mut self, pvh: &PvHeader) -> Result<()>

Add a non-affiliated PV to this VG.

fn new_linear_lv(&mut self, name: &str, extent_size: u64) -> Result<()>

Create a new linear logical volume in the volume group.

fn lv_remove(&mut self, name: &str) -> Result<()>

Destroy a logical volume.